Fix: the cls alias and the git diff tool git alias
The bash cls alias was using escape sequences which worked fine when
it was wrapped in double quotes. But bash convention recommends the
use of single quotes for alias definitions.

I had mad a 'gd' alias for zsh which replaced an existing git diff
alias and replaced it with my custom git difftool alias using icdiff
with 'less -RF' as the PAGER. This also works in bash but I found it
was also possible just to do the same thing in the global .gitconfig
file:
`cmd = 'PAGER="less -RF" icdiff ...`
This enables me to just call `git difftool [file name]` without me
having to remember the `gd` alias.