Fix: many problems within the vimrc to modernize.
There were many issues within the .vimrc that I've been ignoring for
awhile now. I've now updated most of those issues gotten rid of the old
plug-ins and updated it wth the new plun-ins and made the necessary
changes so it should be compatible with `mvim`, Mac OSX `vim` and the
Linux `vim`. But I still need to try it out on Linux to ensure that is a
correct statement.

I also made a number of changes within the .bash_settings file trying to
keep up to date with the .zshrc file. Manily this had to do with
detecting the various flavors and versions of `ls`, `exa`, `lsd` and
`lsg` that many exist on either Linux and Mac OSX. (Quite a pain in the
arse!) Now that this logic is completed I need to actually refactor the
`.zshrc` into a generic file which gets loaded on installation and one
or more files which I can dynamically load from my `settings` repo.