Add: XDG paths, settings for hyper and alacritty
Testing out both terminal UI's alacritty and hyper.  Alacritty is
definately the faster of the two terminals, by far.  But they built the
terminal for this reason alone.  As far as configuration is concerned
the hyper terminal has everything else beat.

I've added a few a few fixes and modifications to the zsh startup
scripts.  One is fixing the pprint functionality which was
misidentifying functions as environment variables.

Adding the user configuration for defining the XDG home directory folder
specification.

Lastly, adding a bare bones motd.ini file for adding the support needed
to add caching inside of the current login_info.py.